crypto/bn256: only use cloudflare for optimized architectures
