+++ crypto/bn256/cloudflare/gfp_decl.go
